+++ CHANGELOG.md
@@ -6,9 +6,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- tonic: Support request timeout.
+
 ### Fixed
 
 - madsim: Replace `SmallRng` with `Xoshiro256PlusPlus` for reproducibility across platforms.
+- etcd: Fix election implementation. Put a key for each candidate.
 
 ## [0.2.17] - 2023-02-17